--- src/main/java/com/kodilla/ecommercee/CartDto.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee;
-
-public class CartDto {
-}
