+++ src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -1,5 +1,8 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
+import com.kodilla.ecommercee.controller.dto.CartDto;
+import com.kodilla.ecommercee.controller.dto.OrderDto;
+import com.kodilla.ecommercee.controller.dto.ProductDto;
 import org.springframework.web.bind.annotation.*;
 import java.util.ArrayList;
 import java.util.List;
