+++ src/main/java/com/kodilla/ecommercee/controller/OrderController.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
 import org.springframework.web.bind.annotation.*;
