+++ src/main/java/com/kodilla/ecommercee/controller/dto/GroupDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.controller.dto;
+
+public class GroupDto {
+}
