+++ src/main/java/com/kodilla/ecommercee/controller/dto/OrderDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.controller.dto;
+
+public class OrderDto {
+}
