+++ src/main/java/com/kodilla/ecommercee/controller/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.controller.dto;
+
+public class ProductDto {
+}
