+++ src/main/java/com/kodilla/ecommercee/util/GenericEntity.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.util;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
